Implementar mais opcodes e trocar uint8_t por uint16_t

+++ chip8.h
@@ -7,12 +7,14 @@ private:
     uint16_t PROGRAM_START_ADDRESS = 0x200;
 
 public:
-    uint8_t gp_regs[16] = { 0 };
+    uint16_t gp_regs[16] = { 0 };
     uint16_t memory[4096] = { 0 };
     uint16_t index_reg = 0;
     uint16_t pc = PROGRAM_START_ADDRESS;
     uint16_t sp;
+
     bool increment_pc_flag = true;
+
     std::stack<uint16_t> stack;
 
     Chip8();
